refactor: remove v2 sub folders
