+++ operate/client/src/modules/components/OperationsPanel/OperationsEntry/index.tsx
