+++ operate/client/src/modules/components/OperationsPanel/index.setup.ts
